backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
aiosqlite>=0.19.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
